Separate `EventHandler` from `BrowserWrapper`
