+++ src/components/content/BaseHtmlContent.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes, { ContentType } from '../prop-types';
 import { PRE_CALCULATION } from '../../constants/CalculationsConstants';
 import Connector from '../../service/connector';
-import { addEventListener } from '../../util/BrowserWrapper';
+import { addEventListener } from '../../util/EventHandler';
 
 export default class BaseHtmlContent extends React.PureComponent {
   constructor(props) {
