+++ src/components/screen/ImageScrollScreen.jsx
@@ -9,10 +9,12 @@ import {
 import {
   scrollTop,
   setScrollTop,
-  addEventListener,
-  removeEventListener,
   waitThenRun,
 } from '../../util/BrowserWrapper';
+import {
+  addEventListener,
+  removeEventListener,
+} from '../../util/EventHandler';
 import PropTypes, { FooterCalculationsType, ContentCalculationsType, ContentType } from '../prop-types';
 import BaseScreen, {
   mapStateToProps as readerBaseScreenMapStateToProps,
